+++ src/js/calculate.js
@@ -42,7 +42,8 @@ export default {
             const dataNodes = dataNodesObj.getElements();
 
             if ( dataNodes.length > 1 ) {
-                if ( updated.repeatPath && name.indexOf( updated.repeatPath ) !== -1 ) {
+
+                if ( updated.repeatPath && name.indexOf( updated.repeatPath + '/' ) !== -1 ) {
                     /*
                      * If the update was triggered by a datanode inside a repeat
                      * and the dependent node is inside the same repeat, we can prevent the expensive index determination