+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Fixed
+- Loading a form with a readonly file input throws an exception.
+
 [5.0.3] - 2019-01-07
 --------------------
 ##### Fixed